refactor(rolldown_plugin_dynamic_import_vars): improve dynamic_import_to_glob
